Moved two vignettes into misc.
