+++ animations/lib/main.dart
@@ -3,6 +3,7 @@
 // found in the LICENSE file.
 
 import 'package:flutter/material.dart';
+
 import 'src/basics/01_animated_container.dart';
 import 'src/basics/02_page_route_builder.dart';
 import 'src/basics/03_animation_controller.dart';
@@ -11,9 +12,9 @@ import 'src/basics/05_animated_builder.dart';
 import 'src/basics/06_custom_tween.dart';
 import 'src/basics/07_tween_sequence.dart';
 import 'src/basics/08_fade_transition.dart';
-import 'src/basics/09_animated_positioned.dart';
-import 'src/basics/10_animated_switcher.dart';
 import 'src/misc/animated_list.dart';
+import 'src/misc/animated_positioned.dart';
+import 'src/misc/animated_switcher.dart';
 import 'src/misc/card_swipe.dart';
 import 'src/misc/carousel.dart';
 import 'src/misc/curved_animation.dart';
@@ -66,14 +67,6 @@ final basicDemos = [
       name: 'Fade Transition',
       route: FadeTransitionDemo.routeName,
       builder: (context) => FadeTransitionDemo()),
-  Demo(
-      name: 'AnimatedPositioned',
-      route: AnimatedPositionedDemo.routeName,
-      builder: (context) => AnimatedPositionedDemo()),
-  Demo(
-      name: 'AnimatedSwitcher',
-      route: AnimatedSwitcherDemo.routeName,
-      builder: (context) => AnimatedSwitcherDemo())
 ];
 
 final miscDemos = [
@@ -105,6 +98,14 @@ final miscDemos = [
       name: 'AnimatedList',
       route: AnimatedListDemo.routeName,
       builder: (context) => AnimatedListDemo()),
+  Demo(
+      name: 'AnimatedPositioned',
+      route: AnimatedPositionedDemo.routeName,
+      builder: (context) => AnimatedPositionedDemo()),
+  Demo(
+      name: 'AnimatedSwitcher',
+      route: AnimatedSwitcherDemo.routeName,
+      builder: (context) => AnimatedSwitcherDemo()),
   Demo(
       name: 'Hero Animation',
       route: HeroAnimationDemo.routeName,
