+++ animations/lib/src/misc/animated_switcher.dart
